1.增加详情页面的上一条下一条功能（咱缺少下一页加载服务端数据的等待效果） 2.删除对话框提示，改为toast

+++ QiQuJoke/QiQuJoke/CommonTableViewController.m
@@ -73,7 +73,7 @@
     else{
         //上拉加载更多数据（从服务器端请求更多数据）
         NSInteger indexOfNeedLoad = _cm.itemsArr.count/kPageDefaultCount;
-        NSLog(@"%d",indexOfNeedLoad);
+        NSLog(@"%ld",(long)indexOfNeedLoad);
         [self loadDataByOperMode:RMLoadMore pIndex:indexOfNeedLoad];
     }
